Project arch refactor
