+++ src/Components/Assignment/Assignment.js
@@ -1,0 +1,13 @@
+
+export class Assignment {
+
+	constructor({ score }) {
+		this._score = score
+	}
+
+
+	get score() {
+		return this._score
+	}
+
+}
